Initial refactor for new project
