+++ .github/workflows/CI.yml
@@ -14,7 +14,7 @@ jobs:
         uses: actions/checkout@v2
 
       - name: Build image
-        run: docker build -f Dockerfile.build -t spring-boot-base-api .
+        run: docker build -f Dockerfile.build -t qotsa-api .
 
       - name: Run tests
-        run: docker run spring-boot-base-api:latest mvn clean verify
+        run: docker run qotsa-api:latest mvn clean verify
