+++ pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.base</groupId>
-    <artifactId>Spring-Boot-Base-API</artifactId>
+    <groupId>com.qotsa.api</groupId>
+    <artifactId>QOTSA-API</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
@@ -27,7 +27,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>com.base.Main</mainClass>
+                            <mainClass>com.qotsa.api.Main</mainClass>
                         </manifest>
                     </archive>
                     <descriptorRefs>
