+++ src/it/java/com/qotsa/api/controller/HelloWorldControllerIT.java
@@ -1,12 +1,10 @@
-package com.base.controller;
+package com.qotsa.api.controller;
 
-import com.base.configuration.TestConfig;
-import com.base.repository.MyDatabaseRepository;
+import com.qotsa.api.configuration.TestConfig;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
