+++ src/main/java/com/qotsa/api/Main.java
@@ -1,4 +1,4 @@
-package com.base;
+package com.qotsa.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
