+++ src/main/java/com/qotsa/api/configuration/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.base.configuration;
+package com.qotsa.api.configuration;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
