+++ src/main/java/com/qotsa/api/controller/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.base.controller;
+package com.qotsa.api.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
