+++ src/main/java/com/qotsa/api/dto/HelloDto.java
@@ -1,4 +1,4 @@
-package com.base.dto;
+package com.qotsa.api.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.persistence.*;
